Feature List
- All in One Dev-Conatiner Image
- k8s mode improvements
- Applicatoin Insight data point improvments

Related work items: #1390, #1703, #1756, #1770, #1783, #1784, #1792, #1793, #1795, #1831, #1834, #1836, #1837

+++ cli/cmd/remove/k8s.go
@@ -1,35 +1,48 @@
 /*
- * @Date: 2022-06-07 14:02:38
- * @LastEditors: Jason Chen
- * @LastEditTime: 2022-07-07 13:07:44
- * @FilePath: /smartide-cli/cmd/remove/k8s.go
- */
-/*
- * @Date: 2022-06-07 14:02:29
- * @LastEditors: Jason Chen
- * @LastEditTime: 2022-06-07 14:02:30
- * @FilePath: /smartide-cli/cmd/remove/local.go
- */
+SmartIDE - Dev Containers
+Copyright (C) 2023 leansoftX.com
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
 
 package remove
 
 import (
+	"fmt"
 	"os"
+	"strings"
 
 	"github.com/leansoftX/smartide-cli/internal/biz/workspace"
 	"github.com/leansoftX/smartide-cli/pkg/common"
-	"github.com/leansoftX/smartide-cli/pkg/kubectl"
+	"github.com/leansoftX/smartide-cli/pkg/k8s"
 	"github.com/spf13/cobra"
 )
 
 // 删除k8s资源
-func RemoveK8s(k8sUtil kubectl.KubernetesUtil, workspaceInfo workspace.WorkspaceInfo) error {
+func RemoveK8s(k8sUtil k8s.KubernetesUtil, workspaceInfo workspace.WorkspaceInfo) error {
 
 	// 移除k8s资源
 	common.SmartIDELog.Info("移除k8s资源...")
 
 	output, err := k8sUtil.ExecKubectlCommandCombined("delete namespace --force "+workspaceInfo.K8sInfo.Namespace, "")
 	common.SmartIDELog.Debug(output)
+	notFoundMsg := "Error from server (NotFound)"
+	if strings.Contains(output, notFoundMsg) {
+		msg := fmt.Sprintf("%v: namespaces \"%v\" not found", notFoundMsg, workspaceInfo.K8sInfo.Namespace)
+		common.SmartIDELog.Importance(msg)
+		return nil
+	}
 	if err != nil {
 		return err
 	}
@@ -41,14 +54,19 @@ func RemoveK8s(k8sUtil kubectl.KubernetesUtil, workspaceInfo workspace.Workspace
 		return err
 	}
 	repoName := common.GetRepoName(workspaceInfo.GitCloneRepoUrl)
-	filePath := common.PathJoin(home, ".ide", repoName)
-	os.RemoveAll(filePath)
+	if repoName != "" {
+		filePath := common.PathJoin(home, ".ide", repoName)
+		os.RemoveAll(filePath)
+	}
+
+	//remove config note from .ssh/config file
+	workspaceInfo.RemoveSSHConfig()
 
 	return nil
 }
 
 // 删除远程工作区对应的k8s
-func RemoveServerK8s(k8sUtil kubectl.KubernetesUtil,
+func RemoveServerK8s(k8sUtil k8s.KubernetesUtil,
 	cmd *cobra.Command, workspaceInfo workspace.WorkspaceInfo,
 	isRemoveAllComposeImages bool, isForce bool,
 	podName string) error {
